Se abstrajo la logica de controller del router de Auth
