--- src/en.json
@@ -1,3 +1,0 @@
-{
-	"greeting.formal": "greeting.formal"
-}
